requirements.txt
Django==1.10
factory-boy==2.7.0